BRAIN- combining integration test files
