+++ README.md
@@ -18,7 +18,7 @@ We want to measure the following values in you:
 The technical portion we will be measuring you on requires you to do the work below. It entails these 3 steps:
 
 0) Setting up the project,
-1) Defining the acceptance criteria,
+1) Give the problem: defining the acceptance criteria,
 2) Iterating on the code base, and
 3) Trying to pass the acceptance criteria.
 
@@ -50,11 +50,10 @@ $ (cd docker && docker-compose up -d)
 
 This will launch the services specified in the [docker-compose.yml](./docker/docker-compose.yml) file
 
-
 | Service          | Name                        | Address           | Description                                                          |
 |------------------|-----------------------------|-------------------|----------------------------------------------------------------------|
 | Adminer          | docker_adminer_1            | `127.0.0.1:11000` | [Adminer](https://www.adminer.org/) instance for inspecting database contents (http) |
-| Mysql         | docker_mysql-database_1  | `127.0.0.1:11001` | A Mysql instance, username is `root`      |
+| Mysql            | docker_mysql-database_1  | `127.0.0.1:11001` | A Mysql instance, username is `root`      |
 | Memcached        | docker_memcached_1          | `127.0.0.1:12002` | A memcached instance                                                 |
 | Redis            | gb-service_redis_1          | `127.0.0.1:12003` | A redis instance                                                     |
 
@@ -63,6 +62,11 @@ To stop the services you can run:
 $ (cd docker && docker-compose down)
 ```
 
-Choose either:
+## Problem
+
+We want to build a webhook microservice that takes a URL and some data, and returns X.
+
+## Choose a framework to solve the problem with
 
 [Node/JS](https://github.com/greenbits/captains-cake/tree/master/node-express) or (Sinatra/Ruby)[https://github.com/greenbits/captains-cake/tree/master/ruby-sinatra]
+
